e2etest: move scylla_proxy_cluster from validator-tests
This commit moves interface of scylla_proxy_cluster actor from validator-test
into e2etest-scylla_proxy_cluster. It also updates crate references.
