+++ crates/validator/src/common.rs
@@ -8,6 +8,8 @@ use async_backtrace::framed;
 use e2etest_dns::DnsExt;
 use e2etest_scylla_cluster::ScyllaClusterExt;
 use e2etest_scylla_cluster::ScyllaNodeConfig;
+use e2etest_scylla_proxy_cluster::ScyllaProxyClusterExt;
+use e2etest_scylla_proxy_cluster::ScyllaProxyNodeConfig;
 use httpclient::HttpClient;
 use itertools::Itertools;
 use scylla::client::session::Session;
@@ -25,8 +27,6 @@ use std::time::Duration;
 use tap::Pipe;
 use tokio::time;
 use tracing::info;
-use vector_search_validator_tests::ScyllaProxyClusterExt;
-use vector_search_validator_tests::ScyllaProxyNodeConfig;
 use vector_search_validator_tests::TlsExt;
 use vector_search_validator_tests::VectorStoreClusterExt;
 use vector_search_validator_tests::VectorStoreNodeConfig;
