+++ crates/validator/src/db_timeout.rs
@@ -6,6 +6,7 @@
 use crate::TestActors;
 use crate::common::*;
 use async_backtrace::framed;
+use e2etest_scylla_proxy_cluster::ScyllaProxyClusterExt;
 use scylla_proxy::Condition;
 use scylla_proxy::Reaction;
 use scylla_proxy::RequestReaction;
@@ -16,7 +17,6 @@ use tokio::sync::mpsc;
 use tokio::sync::watch;
 use tokio::time;
 use tracing::info;
-use vector_search_validator_tests::ScyllaProxyClusterExt;
 use vector_search_validator_tests::TestCase;
 
 #[framed]
