+++ crates/validator/src/lib.rs
@@ -26,6 +26,7 @@ use e2etest_dns::DnsExt;
 use e2etest_firewall::Firewall;
 use e2etest_scylla_cluster::ScyllaCluster;
 use e2etest_scylla_cluster::ScyllaClusterExt;
+use e2etest_scylla_proxy_cluster::ScyllaProxyCluster;
 use std::net::Ipv4Addr;
 use std::path::PathBuf;
 use std::sync::Arc;
@@ -36,7 +37,6 @@ use tracing_subscriber::EnvFilter;
 use tracing_subscriber::filter;
 use tracing_subscriber::fmt;
 use tracing_subscriber::prelude::*;
-use vector_search_validator_tests::ScyllaProxyCluster;
 use vector_search_validator_tests::TestCase;
 use vector_search_validator_tests::Tls;
 use vector_search_validator_tests::VectorStoreCluster;
